Reworked from callback to promise + stability impr
Initially the paradigm used callbacks for the js to native communication, these callbacks were promisified.
After learning we have the option of actually using promises straightaway and letting react native handle the promisification for us, I've reworked the code to account for that.

Cleaned up the error handling, in particular for bluetooth unavailability and pairing failures. It remains to be seen whether the UX for the BLE unavailable case is good enough and if the transport is resilient to these kind of cases.